# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/auth/me:
        get:
            tags:
                - AuthService
            description: |-
                GetCurrentUser returns the authenticated user's information.
                 Validate the access token and returns user details.
                 Similar to OIDC's /userinfo endpoint.
            operationId: AuthService_GetCurrentUser
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetCurrentUserResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/refresh:
        post:
            tags:
                - AuthService
            description: |-
                RefreshToken exchanges a valid refresh token for a new access token.
                 The refresh token is read from the HttpOnly cookie.
                 Returns a new short-lived access token.
            operationId: AuthService_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RefreshTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RefreshTokenResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/signin:
        post:
            tags:
                - AuthService
            description: |-
                SignIn authenticates a user with credentials and returns tokens.
                 On success, returns an access token and sets a refresh token cookie.
                 Supports password-based and SSO authentication methods.
            operationId: AuthService_SignIn
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SignInRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SignInResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/auth/signout:
        post:
            tags:
                - AuthService
            description: |-
                SignOut terminates the user's authentication.
                 Revokes the refresh token and clear the authentication cookie.
            operationId: AuthService_SignOut
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GetCurrentUserResponse:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/User'
                    description: The authenticated user's information.
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        RefreshTokenRequest:
            type: object
            properties: {}
        RefreshTokenResponse:
            type: object
            properties:
                accessToken:
                    type: string
                    description: The new short-lived access token.
                expiresAt:
                    type: string
                    description: When the access token expires.
                    format: date-time
...
        SignInRequest:
            type: object
            properties:
                passwordCredentials:
                    allOf:
                        - $ref: '#/components/schemas/SignInRequest_PasswordCredentials'
                    description: Username and password authentication.
                ssoCredentials:
                    allOf:
                        - $ref: '#/components/schemas/SignInRequest_SSOCredentials'
                    description: SSO provider authentication.
        SignInRequest_PasswordCredentials:
            required:
                - username
                - password
            type: object
            properties:
                username:
                    type: string
                    description: The username to sign in with.
                password:
                    type: string
                    description: The password to sign in with.
            description: Nested message for password-based authentication credentials.
        SignInRequest_SSOCredentials:
            required:
                - idpId
                - code
                - redirectUri
            type: object
            properties:
                idpId:
                    type: integer
                    description: The ID of the SSO provider.
                    format: int32
                code:
                    type: string
                    description: The authentication code from the SSO provider.
                redirectUri:
                    type: string
                    description: The redirect URI used in the SSO flow.
                codeVerifier:
                    type: string
                    description: |-
                        The PKCE code verifier for enhanced security (RFC 7636).
                         Optional - enables PKCE flow protection against authorization code interception.
            description: Nested message for SSO authentication credentials.
        SignInResponse:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/User'
                    description: The authenticated user's information.
                accessToken:
                    type: string
                    description: |-
                        The short-lived access token for API requests.
                         Store in memory only, not in localStorage.
                accessTokenExpiresAt:
                    type: string
                    description: |-
                        When the access token expires.
                         Client should call RefreshToken before this time.
                    format: date-time
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        User:
            required:
                - role
                - username
                - state
            type: object
            properties:
                name:
                    type: string
                    description: |-
                        The resource name of the user.
                         Format: users/{user}
                role:
                    enum:
                        - ROLE_UNSPECIFIED
                        - ADMIN
                        - USER
                    type: string
                    description: The role of the user.
                    format: enum
                username:
                    type: string
                    description: Required. The unique username for login.
                email:
                    type: string
                    description: Optional. The email address of the user.
                displayName:
                    type: string
                    description: Optional. The display name of the user.
                avatarUrl:
                    type: string
                    description: Optional. The avatar URL of the user.
                description:
                    type: string
                    description: Optional. The description of the user.
                password:
                    writeOnly: true
                    type: string
                    description: Input only. The password for the user.
                state:
                    enum:
                        - STATE_UNSPECIFIED
                        - NORMAL
                        - ARCHIVED
                    type: string
                    description: The state of the user.
                    format: enum
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    description: Output only. The last update timestamp.
                    format: date-time
...